+++ public/javascripts/Link.js
@@ -11,6 +11,8 @@
     source.dispatchEvent('link', this);
     target.dispatchEvent('linked', this);
 
+    this.hideInfoPanel = this.hideInfoPanel.bind(this);
+
     if (wp.initialized) {
       source.validate();
       target.updateDownstreams();
@@ -19,6 +21,7 @@
   };
 
   Link.prototype.destroy = function () {
+    this.hideInfoPanel();
     this.source.out.delete(this.target);
     this.target.in.delete(this.source);
     this.source.dispatchEvent('unlink', this.target);
@@ -34,16 +37,52 @@
     delete this.process;
   };
 
-  Link.prototype.showInfoPanel = function (ev) {
-    console.log('show link info: ', this);
+  Link.prototype.showInfoPanel = function (x, y) {
+    var
+    self = this,
+    overlay = this.process.overlay,
+    panel = this.panel || (this.panel = this.buildInfoPanel()),
+    previous = overlay.$('.link-info');
+
+    if (previous) previous.wpobj.hideInfoPanel();
+
+    overlay.grab(panel);
+    panel.style.left = Math.trunc(x - panel.offsetWidth / 2) + 'px';
+    panel.style.top = Math.trunc(y - panel.$('.link-title').offsetHeight) + 'px';
+
+    panel.addEventListener('mousemove', function onmove() {
+      document.addEventListener('click', self.hideInfoPanel);
+      panel.removeEventListener('mousemove', onmove);
+    });
   };
 
   Link.prototype.hideInfoPanel = function () {
-
+    this.panel.unload();
+    document.removeEventListener('click', this.hideInfoPanel);
   };
 
   Link.prototype.buildInfoPanel = function () {
+    var self = this;
 
+    return new Element('div', {
+      class: 'link-info',
+      properties: { wpobj: this }
+    }).adopt(
+      new Element('h5', {
+        class: 'link-title',
+        text: this.source.type.displayName + ' > ' + this.target.type.displayName
+      }),
+
+      new Element('div', {
+        class: 'link-actions'
+      }).adopt(
+        new Element('button', {
+          class: 'link-delete',
+          text: 'delete',
+          events: { click: function () { self.destroy(); } }
+        })
+      )
+    );
   };
 
   Evented(Link);